Move ctxRef marker primitives to a backend-neutral hipthrusts/ctx-ref subpath
The ctxRef marker family is backend-agnostic (it depends only on core
NestedPathReq) but lived in src/mongoose.ts for historical reasons — so the
only way to reach it was the hipthrusts/mongoose entrypoint, which also
declares a mongoose peer dependency. That undercut the whole point of
exposing isCtxRef/SpecReq: letting a NON-mongoose backend emit and recognize
the same markers.

Give the primitives their own home:

- New module src/ctx-ref.ts + subpath hipthrusts/ctx-ref, exporting ctxRef,
  isCtxRef, CtxRef, CtxRefReq, and SpecReq. Backend-neutral; no mongoose peer
  implied. The marker is still keyed by Symbol.for('hipthrusts.ctxRef'), so
  refs stay recognizable across module copies.
- src/mongoose.ts imports the primitives from ./ctx-ref.js and re-exports the
  three names it shipped in 1.0.0 (ctxRef, CtxRef, CtxRefReq) for backward
  compatibility — existing imports from hipthrusts/mongoose keep working.
  isCtxRef and SpecReq were never released from mongoose, so they are NOT
  re-exported there; they live only on the new subpath.
- readCtxPath (a runtime resolver detail) stays private in mongoose.ts.

- package.json: register the ./ctx-ref export (exports map regenerated by tshy)
- tests: dedicated test/ctx-ref.test.ts (guard) + test/ctx-ref.test-d.ts
  (CtxRef/CtxRefReq/SpecReq types); reverted the temporary additions to the
  mongoose test files
- README: note pointing loader authors to hipthrusts/ctx-ref
- CHANGELOG Added/Changed; ROADMAP item updated

Verified: 237 tests, lint, typecheck, format, smoke, and check:exports
(hipthrusts/ctx-ref resolves 🟢 for node16 CJS+ESM and bundler) all pass.
